+++ fourtoon-cookie/src/hooks/account.ts
@@ -3,7 +3,7 @@ import { AccountStatus } from "../types/account";
 import { JWTToken } from "../types/jwt";
 import { Member } from "../types/member";
 import { useCreateMember, useDeleteMember, useMember } from "./server/member";
-import { useJwtStore } from "../store/jwt";
+import { useJwtStore } from "./store/jwt";
 
 
 export const useAccountState = () => {
