+++ fourtoon-cookie/src/hooks/server/member.ts
@@ -2,7 +2,7 @@ import { useMutation, useQuery, useQueryClient } from "react-query";
 import { deleteMember, getMember, postMember } from "../../apis/member"
 import { Member } from "../../types/member";
 import { JwtError } from "../../error/JwtError";
-import { useJwtStore } from "../../store/jwt";
+import { useJwtStore } from "../store/jwt";
 
 export const useMember = () => {
     const queryClient = useQueryClient();
